client: fix bug with safari hanging indefinitely
this was due to safari's behavior regarding preloading of audio/video, i.e.,
safari forbids preloading, which we
however required to determine if we want to show the player to the user
yet. This may cause issues with starting to play tracks. Especially, seeking
of unplayed tracks does not work, because they are not loaded yet.