:bug: watchdog: catch deletion and stop node from going to internal catch
Essentially, we can leave out the catch block if we do not want to perform any
specific treating of the error. As we know what's going wrong (i.e. reading from
a file that has been deleted), we left it out previously. NodeJS however catches the exception
on a global level and does not execute the finally branch afterwards,
as the global catch does not preceed any finally block itself